test: reorganize test folder structure per Rust best practices
- Delete 17 redundant wrapper .rs files that only contained `mod xxx;`
- Create tests/common/ directory for shared utilities (fixtures, macros, assertions)
- Split property_tests.rs (1,157 lines) into focused modules:
  - spacing.rs, css_parsing.rs, buffer_ops.rs, widget_constraints.rs
  - stack_layout.rs, splitter.rs, state_transition.rs
  - saturating_arithmetic.rs, float_conversion.rs
- Split layout_tests.rs (567 lines) into focused modules:
  - merge.rs, constraints.rs, breakpoints.rs
- Fix missing flex_edge_cases.rs in layout_module/
- All tests passing (13 passed, 0 failed, 455 ignored)
