--- tests/boundary_tests.rs
@@ -1,5 +1,0 @@
-//! Boundary condition test entry point
-//!
-//! Aggregates all boundary condition tests.
-
-mod boundary;
