--- tests/canvas_tests.rs
@@ -1,3 +1,0 @@
-//! Tests for Canvas widget - split into modules
-
-mod canvas;
