--- tests/code_editor_tests.rs
@@ -1,3 +1,0 @@
-//! Tests for CodeEditor widget - split into modules
-
-mod code_editor;
