+++ tests/common/mod.rs
@@ -1,0 +1,8 @@
+//! Common test utilities and helpers
+//!
+//! This module provides shared utilities, fixtures, macros, and assertions
+//! used across multiple test modules.
+
+pub mod fixtures;
+pub mod macros;
+pub mod assertions;
