--- tests/data_viewer_tests.rs
@@ -1,3 +1,0 @@
-//! Tests for CSV and JSON viewer widgets - split into modules
-
-mod data_viewer;
