--- tests/event_tests.rs
@@ -1,3 +1,0 @@
-//! Event integration tests
-
-mod event;
