+++ tests/layout_module/mod.rs
@@ -1,6 +1,9 @@
 //! Integration tests for the layout module - split into modules
 
+mod breakpoints;
+mod constraints;
 mod engine;
 mod flex_edge_cases;
+mod merge;
 mod rect;
 mod responsive;
