--- tests/log_viewer_tests.rs
@@ -1,3 +1,0 @@
-//! Tests for LogViewer widget - split into modules
-
-mod log_viewer;
