+++ tests/property/mod.rs
@@ -1,7 +1,16 @@
 //! Property-based tests for Revue core components - split into modules
 
+mod buffer_ops;
 mod color;
+mod css_parsing;
+mod float_conversion;
 mod layout_edge;
 mod rect;
+mod saturating_arithmetic;
 mod signal;
+mod spacing;
+mod splitter;
+mod stack_layout;
+mod state_transition;
 mod string_text;
+mod widget_constraints;
