--- tests/reactive_edge_cases.rs
@@ -1,3 +1,0 @@
-//! Edge case tests for the reactive system - split into modules
-
-mod reactive_edge;
