--- tests/reactive_tests.rs
@@ -1,3 +1,0 @@
-//! Reactive integration tests
-
-mod reactive;
