--- tests/render_tests.rs
@@ -1,3 +1,0 @@
-//! Integration tests for the render module - split into modules
-
-mod render_module;
