--- tests/rich_text_editor_tests.rs
@@ -1,3 +1,0 @@
-//! Tests for RichTextEditor widget - split into modules
-
-mod rich_text_editor;
