--- tests/sidebar_tests.rs
@@ -1,3 +1,0 @@
-//! Tests for Sidebar Navigation widget - split into modules
-
-mod sidebar;
