--- tests/style_tests.rs
@@ -1,3 +1,0 @@
-//! Style integration tests
-
-mod style;
