--- tests/widget_edge_cases.rs
@@ -1,7 +1,0 @@
-//! Widget edge case test entry point
-//!
-//! Aggregates all widget edge case tests.
-
-mod widget {
-    pub mod edge_cases;
-}
