--- tests/widget_snapshots.rs
@@ -1,7 +1,0 @@
-//! Widget snapshot tests - split into modules by widget category
-//!
-//! Tests that verify widget rendering output matches expected snapshots.
-//! Run with: cargo test --test widget_snapshots
-//! Update snapshots: REVUE_UPDATE_SNAPSHOTS=1 cargo test
-
-mod snapshots_widget;
