--- tests/widget_tests.rs
@@ -1,6 +1,0 @@
-//! Widget integration tests
-//!
-//! Tests for widgets that use only public API.
-//! Tests accessing private fields remain inline in src/widget/*.rs.
-
-mod widget;
