--- tests/worker_tests.rs
@@ -1,5 +1,0 @@
-//! Worker system integration tests
-//!
-//! Tests for worker pool, task handles, and channel communication.
-
-mod worker;
